+++ app.js
@@ -6,6 +6,5 @@ module.exports = app => {
   }
 
   app.beforeStart(async () => {
-    app.mongodbClient;
   });
 };
