+++ app/common/base_model.js
@@ -4,7 +4,7 @@
  * 给所有的 Model 扩展功能
  * http://mongoosejs.com/docs/plugins.html
  */
-const tools = require('../common/tools');
+const tools = require('./tools');
 
 module.exports = function(schema) {
   schema.methods.create_at_ago = function() {
