+++ app/service/topic.js
@@ -1,212 +1,171 @@
 'use strict';
 
-const mongoose = require('mongoose');
-const BaseModel = require('../models/base_model');
-const Schema = mongoose.Schema;
-const ObjectId = Schema.ObjectId;
+const Service = require('egg').Service;
 
-module.exports = app => {
-
-  const config = app.config;
-
-  const TopicSchema = new Schema({
-    title: { type: String },
-    content: { type: String },
-    author_id: { type: ObjectId },
-    top: { type: Boolean, default: false }, // 置顶帖
-    good: { type: Boolean, default: false }, // 精华帖
-    lock: { type: Boolean, default: false }, // 被锁定主题
-    reply_count: { type: Number, default: 0 },
-    visit_count: { type: Number, default: 0 },
-    collect_count: { type: Number, default: 0 },
-    create_at: { type: Date, default: Date.now },
-    update_at: { type: Date, default: Date.now },
-    last_reply: { type: ObjectId },
-    last_reply_at: { type: Date, default: Date.now },
-    content_is_html: { type: Boolean },
-    tab: { type: String },
-    deleted: { type: Boolean, default: false },
-  });
-
-  TopicSchema.plugin(BaseModel);
-  TopicSchema.index({ create_at: -1 });
-  TopicSchema.index({ top: -1, last_reply_at: -1 });
-  TopicSchema.index({ author_id: 1, create_at: -1 });
-
-  TopicSchema.virtual('tabName').get(function() {
-    const tab = this.tab;
-    const pair = config.tabs.find(function(_pair) {
-      return _pair[0] === tab;
-    });
-
-    if (pair) {
-      return pair[1];
-    }
-    return '';
-  });
-
-  const Topic = mongoose.model('Topic', TopicSchema);
-
-  return class Service extends app.Service {
-
-    /*
-     * 根据主题ID获取主题
-     * @param {String} id 主题ID
-     */
-    async getTopicById(id) {
-      const topic = await Topic.findOne({ _id: id }).exec();
-      if (!topic) {
-        return {
-          topic: null,
-          author: null,
-          last_reply: null,
-        };
-      }
-
-      const author = await this.service.user.getUserById(topic.author_id);
-
-      let last_reply = null;
-      if (topic.last_reply) {
-        last_reply = await this.service.reply.getReplyById(topic.last_reply);
-      }
-
+class TopicService extends Service {
+  /*
+   * 根据主题ID获取主题
+   * @param {String} id 主题ID
+   */
+  async getTopicById(id) {
+    const topic = await this.ctx.model.Topic.findOne({ _id: id }).exec();
+    if (!topic) {
       return {
-        topic,
-        author,
-        last_reply,
+        topic: null,
+        author: null,
+        last_reply: null,
       };
     }
 
-    /*
-     * 获取关键词能搜索到的主题数量
-     * @param {String} query 搜索关键词
-     */
-    getCountByQuery(query) {
-      return Topic.count(query).exec();
+    const author = await this.service.user.getUserById(topic.author_id);
+
+    let last_reply = null;
+    if (topic.last_reply) {
+      last_reply = await this.service.reply.getReplyById(topic.last_reply);
     }
 
-    /*
-     * 根据关键词，获取主题列表
-     * @param {String} query 搜索关键词
-     * @param {Object} opt 搜索选项
-     */
-    async getTopicsByQuery(query, opt) {
-      query.deleted = false;
-      const topics = await Topic.find(query, {}, opt).exec();
+    return {
+      topic,
+      author,
+      last_reply,
+    };
+  }
 
-      if (topics.length === 0) {
-        return [];
-      }
+  /*
+   * 获取关键词能搜索到的主题数量
+   * @param {String} query 搜索关键词
+   */
+  getCountByQuery(query) {
+    return this.ctx.model.Topic.count(query).exec();
+  }
 
-      await Promise.all(topics.map(async topic => {
-        const { author, reply } = await Promise.all([
-          this.service.user.getUserById(topic.author_id),
-          // 获取主题的最后回复
-          this.service.reply.getReplyById(topic.last_reply),
-        ]);
-        topic.author = author;
-        topic.reply = reply;
-      }));
+  /*
+   * 根据关键词，获取主题列表
+   * @param {String} query 搜索关键词
+   * @param {Object} opt 搜索选项
+   */
+  async getTopicsByQuery(query, opt) {
+    query.deleted = false;
+    const topics = await this.ctx.model.Topic.find(query, {}, opt).exec();
 
-      return topics.filter(item => {
-        // 删除不合规的 topic
-        return !!item.author;
-      });
+    if (topics.length === 0) {
+      return [];
     }
 
-    // for sitemap
-    getLimit5w() {
-      const query = { deleted: false };
-      const opts = { limit: 50000, sort: '-create_at' };
-      return Topic.find(query, '_id', opts).exec();
+    await Promise.all(topics.map(async topic => {
+      const { author, reply } = await Promise.all([
+        this.service.user.getUserById(topic.author_id),
+        // 获取主题的最后回复
+        this.service.reply.getReplyById(topic.last_reply),
+      ]);
+      topic.author = author;
+      topic.reply = reply;
+    }));
+
+    return topics.filter(item => {
+      // 删除不合规的 topic
+      return !!item.author;
+    });
+  }
+
+  // for sitemap
+  getLimit5w() {
+    const query = { deleted: false };
+    const opts = { limit: 50000, sort: '-create_at' };
+    return this.ctx.model.Topic.find(query, '_id', opts).exec();
+  }
+
+  /*
+   * 获取所有信息的主题
+   * Callback:
+   * - err, 数据库异常
+   * - message, 消息
+   * - topic, 主题
+   * - author, 主题作者
+   * - replies, 主题的回复
+   * @param {String} id 主题ID
+   * @param {Function} callback 回调函数
+   */
+  async getFullTopic(id) {
+    const query = { _id: id, deleted: false };
+    const topic = await this.ctx.model.Topic.findOne(query);
+
+    if (!topic) {
+      return [ '此话题不存在或已被删除。' ];
     }
 
-    /*
-     * 获取所有信息的主题
-     * Callback:
-     * - err, 数据库异常
-     * - message, 消息
-     * - topic, 主题
-     * - author, 主题作者
-     * - replies, 主题的回复
-     * @param {String} id 主题ID
-     * @param {Function} callback 回调函数
-     */
-    async getFullTopic(id) {
-      const query = { _id: id, deleted: false };
-      const topic = await Topic.findOne(query);
+    topic.linkedContent = this.service.at.linkUsers(topic.content);
 
-      if (!topic) {
-        return [ '此话题不存在或已被删除。' ];
-      }
-
-      topic.linkedContent = this.service.at.linkUsers(topic.content);
-
-      const author = await this.service.user.getUserById(topic.author_id);
-      if (!author) {
-        return [ '话题的作者丢了。' ];
-      }
-
-      const replies = await this.service.reply.getRepliesByTopicId(topic._id);
-      return [ '', topic, author, replies ];
+    const author = await this.service.user.getUserById(topic.author_id);
+    if (!author) {
+      return [ '话题的作者丢了。' ];
     }
 
-    /*
-     * 更新主题的最后回复信息
-     * @param {String} topicId 主题ID
-     * @param {String} replyId 回复ID
-     * @param {Function} callback 回调函数
-     */
-    async updateLastReply(topicId, replyId) {
-      const topic = await Topic.findOne({ _id: topicId }).exec();
-      if (!topic) {
-        return;
-      }
-      topic.last_reply = replyId;
-      topic.last_reply_at = new Date();
-      topic.reply_count += 1;
-      return topic.save();
+    const replies = await this.service.reply.getRepliesByTopicId(topic._id);
+    return [
+      '',
+      topic,
+      author,
+      replies,
+    ];
+  }
+
+  /*
+   * 更新主题的最后回复信息
+   * @param {String} topicId 主题ID
+   * @param {String} replyId 回复ID
+   * @param {Function} callback 回调函数
+   */
+  async updateLastReply(topicId, replyId) {
+    const topic = await this.ctx.model.Topic.findOne({ _id: topicId }).exec();
+    if (!topic) {
+      return;
+    }
+    topic.last_reply = replyId;
+    topic.last_reply_at = new Date();
+    topic.reply_count += 1;
+    return topic.save();
+  }
+
+  /*
+   * 根据主题ID，查找一条主题
+   * @param {String} id 主题ID
+   * @param {Function} callback 回调函数
+   */
+  getTopic(id) {
+    return this.ctx.model.Topic.findOne({ _id: id }).exec();
+  }
+
+  /*
+   * 将当前主题的回复计数减1，并且更新最后回复的用户，删除回复时用到
+   * @param {String} id 主题ID
+   */
+  async reduceCount(id) {
+    const topic = await this.ctx.model.Topic.findOne({ _id: id }).exec();
+    if (!topic) {
+      throw new Error('该主题不存在');
     }
 
-    /*
-     * 根据主题ID，查找一条主题
-     * @param {String} id 主题ID
-     * @param {Function} callback 回调函数
-     */
-    getTopic(id) {
-      return Topic.findOne({ _id: id }).exec();
+    topic.reply_count -= 1;
+    const reply = await this.service.reply.getLastReplyByTopId(id);
+    if (reply.length !== 0) {
+      topic.last_reply = reply[0]._id;
+    } else {
+      topic.last_reply = null;
     }
 
-    /*
-     * 将当前主题的回复计数减1，并且更新最后回复的用户，删除回复时用到
-     * @param {String} id 主题ID
-     */
-    async reduceCount(id) {
-      const topic = await Topic.findOne({ _id: id }).exec();
-      if (!topic) {
-        throw new Error('该主题不存在');
-      }
+    return topic.save();
+  }
 
-      topic.reply_count -= 1;
-      const reply = await this.service.reply.getLastReplyByTopId(id);
-      if (reply.length !== 0) {
-        topic.last_reply = reply[0]._id;
-      } else {
-        topic.last_reply = null;
-      }
+  newAndSave(title, content, tab, authorId) {
+    const topic = new this.ctx.model.Topic();
+    topic.title = title;
+    topic.content = content;
+    topic.tab = tab;
+    topic.author_id = authorId;
 
-      return topic.save();
-    }
+    return topic.save();
+  }
+}
 
-    newAndSave(title, content, tab, authorId) {
-      const topic = new Topic();
-      topic.title = title;
-      topic.content = content;
-      topic.tab = tab;
-      topic.author_id = authorId;
-
-      return topic.save();
-    }
-  };
-
-};
+module.exports = TopicService;
