+++ app/service/topic_collect.js
@@ -1,47 +1,32 @@
 'use strict';
 
-const mongoose = require('mongoose');
+const Service = require('egg').Service;
 
-const BaseModel = require('../models/base_model');
-const Schema = mongoose.Schema;
-const ObjectId = Schema.ObjectId;
+class TopicCollectService extends Service {
 
-module.exports = app => {
+  getTopicCollect(userId, topicId) {
+    const query = { user_id: userId, topic_id: topicId };
+    return this.ctx.model.TopicCollect.findOne(query).exec();
+  }
 
-  const TopicCollectSchema = new Schema({
-    user_id: { type: ObjectId },
-    topic_id: { type: ObjectId },
-    create_at: { type: Date, default: Date.now },
-  });
+  getTopicCollectsByUserId(userId, opt) {
+    const defaultOpt = { sort: '-create_at' };
+    opt = Object.assign(defaultOpt, opt);
+    return this.ctx.model.TopicCollect.find({ user_id: userId }, '', opt).exec();
+  }
 
-  TopicCollectSchema.plugin(BaseModel);
-  TopicCollectSchema.index({ user_id: 1, topic_id: 1 }, { unique: true });
+  newAndSave(userId, topicId) {
+    const topic_collect = new this.ctx.model.TopicCollect();
+    topic_collect.user_id = userId;
+    topic_collect.topic_id = topicId;
+    return topic_collect.save();
+  }
 
-  const TopicCollect = mongoose.model('TopicCollect', TopicCollectSchema);
+  remove(userId, topicId) {
+    const query = { user_id: userId, topic_id: topicId };
+    return this.ctx.model.TopicCollect.remove(query).exec();
+  }
+}
 
-  return class Service extends app.Service {
 
-    getTopicCollect(userId, topicId) {
-      const query = { user_id: userId, topic_id: topicId };
-      return TopicCollect.findOne(query).exec();
-    }
-
-    getTopicCollectsByUserId(userId, opt) {
-      const defaultOpt = { sort: '-create_at' };
-      opt = Object.assign(defaultOpt, opt);
-      return TopicCollect.find({ user_id: userId }, '', opt).exec();
-    }
-
-    newAndSave(userId, topicId) {
-      const topic_collect = new TopicCollect();
-      topic_collect.user_id = userId;
-      topic_collect.topic_id = topicId;
-      return topic_collect.save();
-    }
-
-    remove(userId, topicId) {
-      const query = { user_id: userId, topic_id: topicId };
-      return TopicCollect.remove(query).exec();
-    }
-  };
-};
+module.exports = TopicCollectService;
