+++ config/config.local.js
@@ -8,3 +8,13 @@ exports.redis = {
     db: 0,
   },
 };
+
+/**
+ * @see http://mongodb.github.io/node-mongodb-native/2.2/api/Db.html#createCollection
+ */
+exports.mongoose = {
+  url: 'mongodb://127.0.0.1/node_club_test',
+  options: {
+    server: { poolSize: 20 },
+  },
+};
