+++ config/plugin.js
@@ -12,3 +12,8 @@ exports.redis = {
   enable: true,
   package: 'egg-redis',
 };
+
+exports.mongoose = {
+  enable: true,
+  package: 'egg-mongoose',
+};
